scanner/requirements.txt
opencv-python
pandas
pyzbar
Pillow
requests
gspread
oauth2client